// src/serverApi/v3/models/meta-tag-extractor-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Schulcloud-Verbund-Software Server API
 * This is v3 of Schulcloud-Verbund-Software Server. Checkout /docs for v1.
 *
 * The version of the OpenAPI document: 3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { MetaDataEntityType } from './meta-data-entity-type';

/**
 * 
 * @export
 * @interface MetaTagExtractorResponse
 */
export interface MetaTagExtractorResponse {
    /**
     * 
     * @type {string}
     * @memberof MetaTagExtractorResponse
     */
    url: string;
    /**
     * 
     * @type {string}
     * @memberof MetaTagExtractorResponse
     */
    title: string;
    /**
     * 
     * @type {string}
     * @memberof MetaTagExtractorResponse
     */
    description: string;
    /**
     * 
     * @type {string}
     * @memberof MetaTagExtractorResponse
     */
    originalImageUrl?: string;
    /**
     * 
     * @type {string}
     * @memberof MetaTagExtractorResponse
     */
    imageUrl?: string;
    /**
     * 
     * @type {MetaDataEntityType}
     * @memberof MetaTagExtractorResponse
     */
    type: MetaDataEntityType;
    /**
     * 
     * @type {string}
     * @memberof MetaTagExtractorResponse
     */
    parentTitle?: string;
    /**
     * 
     * @type {MetaDataEntityType}
     * @memberof MetaTagExtractorResponse
     */
    parentType?: MetaDataEntityType;
}